requirements.txt
typer[all]>=0.9.0
datasets>=2.14.0
polars>=0.19.0
pyarrow>=13.0.0
regex>=2023.10.0
simhash>=2.1.2
razdel>=0.5.0
xxhash>=3.4.1
transformers>=4.35.0
torch>=2.1.0
safetensors>=0.3.0
numpy>=1.24.0
google-generativeai>=0.3.2
pyyaml>=6.0
rich>=13.7.0
